requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
cycler==0.11.0
discord.py==1.7.3
DiscordUtils==1.3.4
dnspython==2.2.1
docopt==0.6.2
fonttools==4.32.0
icrawler==0.6.6
idna==3.3
kiwisolver==1.4.2
lxml==4.8.0
matplotlib==3.5.1
multidict==6.0.2
num2words==0.5.10
numpy==1.22.3
packaging==21.3
Pillow==9.1.0
pipreqs==0.4.11
praw==7.5.0
prawcore==2.3.0
py-cord==2.0.0b5
pycparser==2.21
pyimgur==0.6.0
pymongo==4.1.0
PyNaCl==1.4.0
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
typing_extensions==4.1.1
update-checker==0.18.0
urllib3==1.26.9
websocket-client==1.3.2
yarg==0.1.9
yarl==1.7.2
youtube-dl==2021.12.17
